+++ R/topojson-choropleth.R
@@ -9,37 +9,12 @@ topoJSONChoroplethDependency <- function() {
   )
 }
 
-#' @param map The leaflet map
-#' @param data Geojson or topojson data: either as a list or a string
-#' @param layerId a unique ID for the layer
-#' @param group the group this layer will be added to
-#' @param valueProperty The property to use for coloring
-#' @param popupProperty The property to use for popup content
-#' @param popupOptions The Options for the popup
-#' @param labelProperty The property to use for labelling.
-#' @param labelOptions The Options for the label
-#' @param scale The scale to use from chroma.js
-#' @param steps number of breakes
-#' @param mode q for quantile, e for equidistant, k for k-means
-#' @param colors overrides scale with manual colors
-#' @param stroke whether to draw stroke along the path (e.g. the borders of
-#'   polygons or circles)
-#' @param color stroke color
-#' @param weight stroke width in pixels
-#' @param opacity stroke opacity (or layer opacity for tile layers)
-#'   circles)
-#' @param fillOpacity fill opacity
-#' @param dashArray a string that defines the stroke
-#'   \href{https://developer.mozilla.org/en/SVG/Attribute/stroke-dasharray}{dash
-#'   pattern}
-#' @param smoothFactor how much to simplify the polyline on each zoom level
-#'   (more means better performance and less accurate representation)
-#' @param noClip whether to disable polyline clipping
-#' @param highlightOptions Options for highlighting shapes on hover.
-#' @rdname topojson-choropleth
+#' Adds a TopoJSON Choropleth.
+#' @param topojson either as a list or a string
+#' @rdname geojson-choropleth
 #' @export
 addTopoJSONChoropleth = function(
-  map, data, layerId = NULL, group = NULL,
+  map, topojson, layerId = NULL, group = NULL,
   valueProperty,
   labelProperty = NULL, labelOptions = leaflet::labelOptions(),
   popupProperty = NULL, popupOptions = leaflet::popupOptions(),
@@ -55,16 +30,14 @@ addTopoJSONChoropleth = function(
   dashArray = NULL,
   smoothFactor = 1.0,
   noClip = FALSE,
+  pathOptions = leaflet::pathOptions(),
   highlightOptions = NULL
 ) {
+  map$dependencies <- c(map$dependencies, omnivoreDependencies())
   map$dependencies <- c(map$dependencies,
                         topoJSONChoroplethDependency())
-  options = list(
+  pathOptions =c(pathOptions, list(
     valueProperty=valueProperty,
-    popupProperty=popupProperty,
-    popupOptions=popupOptions,
-    labelProperty=labelProperty,
-    labelOptions=labelOptions,
     scale=scale,
     steps=steps,
     mode=mode,
@@ -76,24 +49,12 @@ addTopoJSONChoropleth = function(
     fillOpacity=fillOpacity,
     dashArray=dashArray,
     smoothFactor=smoothFactor,
-    noClip=noClip,
-    highlightOptions = highlightOptions
-  )
+    noClip=noClip
+  ))
   leaflet::invokeMethod(
     map, leaflet::getMapData(map), 'addTopoJSONChoropleth',
-    data, layerId, group, options)
-}
-
-#' removes the topojson choropleth.
-#' @rdname topojson-choropleth
-#' @export
-removeTopoJSONChoropleth = function(map, layerId) {
-    leaflet::invokeMethod(map, leaflet::getMapData(map), 'removeTopoJSONChoropleth', layerId)
-}
-
-#' clears the topojson choropleth.
-#' @rdname topojson-choropleth
-#' @export
-clearTopoJSONChoropleth = function(map) {
-    leaflet::invokeMethod(map, NULL, 'clearTopoJSONChoropleth')
+    topojson, layerId, group,
+    labelProperty, labelOptions, popupProperty, popupOptions,
+    pathOptions, highlightOptions
+    )
 }