+++ inst/examples/topojson-choropleth.R
@@ -16,5 +16,5 @@ leaflet() %>%
     color='#ffffff', weight=1, fillOpacity = 0.7,
     highlightOptions =
       highlightOptions(fillOpacity=1, weight=2, opacity=1, color='#000000',
-                        bringToFront=TRUE)
+                        bringToFront=TRUE, sendToBack = TRUE)
   )